なおし

+++ ITP1_10/ITP1_10_A.java
@@ -5,10 +5,10 @@ class Main {
         Scanner scan = new Scanner(System.in);
 
         String[] number = scan.nextLine().split(" ");
-        int x1 = Integer.parseInt(number[0]);
-        int x2 = Integer.parseInt(number[2]);
-        int y1 = Integer.parseInt(number[1]);
-        int y2 = Integer.parseInt(number[3]);
+        double x1 = Double.parseDouble(number[0]);
+        double x2 = Double.parseDouble(number[2]);
+        double y1 = Double.parseDouble(number[1]);
+        double y2 = Double.parseDouble(number[3]);
         double distance = Math.sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
         System.out.println(distance);
     }
